requirements.txt
numpy==1.8.0
nltk==2.0.5
scipy==0.13.2
scikit-learn==0.15.2
goose-extractor==1.0.25
newspaper==0.0.9.8
networkx==1.9.1